refactor(site/src/pages/AgentsPage/components/ChatModelAdminPanel): dedupe Stage 2 story fixtures
